backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
jinja2>=3.1.4
python-multipart>=0.0.9
python-dotenv>=1.0.1
google-adk>=1.2.1
google-genai>=1.0.0
groq>=0.11.0
pytest>=8.3.0
httpx>=0.27.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<4.1
pgvector>=0.3.0
pymupdf>=1.24.0
